+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/design/AbstractNcssCountRule.java
@@ -1,9 +1,10 @@
-/**
+/*
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
 package net.sourceforge.pmd.lang.apex.rule.design;
 
+import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.apex.ast.ASTBreakStatement;
 import net.sourceforge.pmd.lang.apex.ast.ASTContinueStatement;
 import net.sourceforge.pmd.lang.apex.ast.ASTDoLoopStatement;
@@ -27,7 +28,10 @@ import net.sourceforge.pmd.util.NumericConstants;
  * <a href="http://www.kclee.de/clemens/java/javancss/">JavaNCSS rules</a>.
  *
  * @author ported from Java original of Jason Bennett
+ * @deprecated Internal API
  */
+@Deprecated
+@InternalApi
 public abstract class AbstractNcssCountRule extends AbstractStatisticalApexRule {
 
     private Class<?> nodeClass;